DiffusionDPO/requirements.txt
accelerate==0.20.2
diffusers==0.20.0
torchvision
transformers==4.30.2
datasets
ftfy
tensorboard
Jinja2
bitsandbytes
xformers
hpsv2
openai_clip==1.0.1
